backend/requirements.txt
# Django Core
Django==5.2.6
djangorestframework==3.16.0

# Authentication
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# API Documentation
drf-spectacular==0.27.2

# Filtering
django-filter==23.3

# CORS
django-cors-headers==4.3.1

# Environment variables
python-decouple==3.8

# Image processing
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

